requirements-dev.txt
pytest>=8.3.0
scipy>=1.8.0
numpy>=1.20.0
pillow>=9.0.0